requirements.txt
pandas
numpy
scikit-learn
ipykernel
pyYAML
python-dotenv
pymongo
